refactor: Move responder to responders package
Moved the `responder.go` file to a new `responders` package
to better organize the codebase.  Updated imports in
`plugin.go` to reflect this change.
